CLog: fix incorrect use of fprintf()
The first parameter of any *printf function should be should be a
string literal. If you pass in some random string, then the %s, %p etc
that might be in the random string will be interpreted, with whatever
garbage is currently on the stack.

We could now convert this to fwrite(), but just using the most simply
string conversion seemed more simple, this isn't time critical code
anyway.